+++ src/main/java/com/braintreegateway/PartnerMerchant.java
@@ -37,4 +37,3 @@ public final class PartnerMerchant {
         return clientSideEncryptionKey;
     }
 }
-