+++ src/main/java/com/braintreegateway/TransactionOptionsRequest.java
@@ -11,6 +11,8 @@ public class TransactionOptionsRequest extends Request {
     private String venmoSdkSession;
     private String payeeEmail;
     private Boolean skipAdvancedFraudChecking;
+    private Boolean skipAvs;
+    private Boolean skipCvv;
     private TransactionOptionsPayPalRequest transactionOptionsPayPalRequest;
     private TransactionOptionsAmexRewardsRequest transactionOptionsAmexRewardsRequest;
     private TransactionOptionsThreeDSecureRequest transactionOptionsThreeDSecureRequest;
@@ -68,6 +70,16 @@ public class TransactionOptionsRequest extends Request {
         return this;
     }
 
+    public TransactionOptionsRequest skipAvs(Boolean skipAvs) {
+        this.skipAvs = skipAvs;
+        return this;
+    }
+
+    public TransactionOptionsRequest skipCvv(Boolean skipCvv) {
+        this.skipCvv = skipCvv;
+        return this;
+    }
+
     public TransactionOptionsPayPalRequest paypal() {
         transactionOptionsPayPalRequest = new TransactionOptionsPayPalRequest(this);
         return transactionOptionsPayPalRequest;
@@ -109,6 +121,8 @@ public class TransactionOptionsRequest extends Request {
             addElement("venmoSdkSession", venmoSdkSession).
             addElement("payeeEmail", payeeEmail).
             addElement("skipAdvancedFraudChecking", skipAdvancedFraudChecking).
+            addElement("skipAvs", skipAvs).
+            addElement("skipCvv", skipCvv).
             addElement("threeDSecure", transactionOptionsThreeDSecureRequest).
             addElement("paypal", transactionOptionsPayPalRequest).
             addElement("payWithAmexRewards", transactionOptionsAmexRewardsRequest);