+++ src/main/java/com/braintreegateway/WebhookTestingGateway.java
@@ -54,6 +54,8 @@ public class WebhookTestingGateway {
             case PARTNER_MERCHANT_CONNECTED: return partnerMerchantConnectedXml(id);
             case PARTNER_MERCHANT_DISCONNECTED: return partnerMerchantDisconnectedXml(id);
             case PARTNER_MERCHANT_DECLINED: return partnerMerchantDeclinedXml(id);
+            case CONNECTED_MERCHANT_STATUS_TRANSITIONED: return connectedMerchantStatusTransitionedXml(id);
+            case CONNECTED_MERCHANT_PAYPAL_STATUS_CHANGED: return connectedMerchantPayPalStatusChangedXml(id);
             case SUBSCRIPTION_CHARGED_SUCCESSFULLY: return subscriptionChargedSuccessfullyXml(id);
             case ACCOUNT_UPDATER_DAILY_REPORT: return accountUpdaterDailyReportXml(id);
             default: return subscriptionXml(id);
@@ -317,6 +319,22 @@ public class WebhookTestingGateway {
     private String partnerMerchantDeclinedXml(String id) {
         return node("partner-merchant",
                 node("partner-merchant-id", "abc123")
+        );
+    }
+
+    private String connectedMerchantStatusTransitionedXml(String id) {
+        return node("connected-merchant-status-transitioned",
+                node("oauth-application-client-id", "oauth_application_client_id"),
+                node("merchant-public-id", id),
+                node("status", "new_status")
+        );
+    }
+
+    private String connectedMerchantPayPalStatusChangedXml(String id) {
+        return node("connected-merchant-paypal-status-changed",
+                node("oauth-application-client-id", "oauth_application_client_id"),
+                node("merchant-public-id", id),
+                node("action", "link")
         );
     }