+++ src/test/java/com/braintreegateway/integrationtest/PaymentMethodIT.java
@@ -1,5 +1,6 @@
 package com.braintreegateway.integrationtest;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.ArrayList;
@@ -43,18 +44,22 @@ public class PaymentMethodIT extends IntegrationTest {
     }
 
     @Test
-    public void createPayPalAccountFromNonceWithPayeeEmail() {
+    public void createPayPalAccountFromNonceWithPayPalOptions() {
         Result<Customer> customerResult = gateway.customer().create(new CustomerRequest());
         assertTrue(customerResult.isSuccess());
         Customer customer = customerResult.getTarget();
 
-        String nonce = TestHelper.generateFuturePaymentPayPalNonce(gateway);
+        String nonce = TestHelper.generateOrderPaymentPayPalNonce(gateway);
         PaymentMethodRequest request = new PaymentMethodRequest().
             customerId(customer.getId()).
             paymentMethodNonce(nonce).
             options().
                 paypal().
                     payeeEmail("payee@example.com").
+                    orderId("merchant-order-id").
+                    customField("custom merchant field").
+                    description("merchant description").
+                    amount(new BigDecimal("1.23")).
                     done().
                 done();